moved some originInfo date processing to new origin_info.rb - stage one refactoring
