+++ lib/stanford-mods/searchworks.rb
@@ -400,14 +400,7 @@ module Stanford
       def pub_dates
         return dates_marc_encoding unless dates_marc_encoding.empty?
         return dates_no_marc_encoding unless dates_no_marc_encoding.empty?
-        return nil
-      end
-
-      def is_number?(object)
-        true if Integer(object) rescue false
-      end
-      def is_date?(object)
-        true if Date.parse(object) rescue false
+        nil
       end
 
       # Get the publish year from mods
@@ -670,155 +663,6 @@ module Stanford
         @subject_topics ||= self.term_values([:subject, :topic])
       end
 
-      #get a 4 digit year like 1865 from the date array
-      def get_plain_four_digit_year dates
-        dates.each do |f_date|
-          matches=f_date.scan(/\d{4}/)
-          if matches.length == 1
-            @pub_year=matches.first
-            return matches.first
-          else
-            #if there are multiples, check for ones with CE after them
-            matches.each do |match|
-              #look for things like '1865-6 CE'
-              pos = f_date.index(Regexp.new(match+'...CE'))
-              pos = pos ? pos.to_i : 0
-              if f_date.include?(match+' CE') or pos > 0
-                @pub_year=match
-                return match
-              end
-            end
-            return matches.first
-          end
-        end
-        return nil
-      end
-
-      # If a year has a "u" in it, replace instances of u with 0
-      # @param [String] dates
-      # @return String
-      def get_u_year dates
-        dates.each do |f_date|
-          # Single digit u notation
-          matches = f_date.scan(/\d{3}u/)
-          if matches.length == 1
-            return matches.first.gsub('u','0')
-          end
-          # Double digit u notation
-          matches = f_date.scan(/\d{2}u{2}/)
-          if matches.length == 1
-            return matches.first.gsub('u','-')
-          end
-        end
-        return nil
-      end
-
-      #get a double digit century like '12th century' from the date array
-      def get_double_digit_century dates
-        dates.each do |f_date|
-          matches=f_date.scan(/\d{2}th/)
-          if matches.length == 1
-            @pub_year=((matches.first[0,2].to_i)-1).to_s+'--'
-            return @pub_year
-          end
-          #if there are multiples, check for ones with CE after them
-          if matches.length > 0
-            matches.each do |match|
-              pos = f_date.index(Regexp.new(match+'...CE'))
-              pos = pos ? pos.to_i : f_date.index(Regexp.new(match+' century CE'))
-              pos = pos ? pos.to_i : 0
-              if f_date.include?(match+' CE') or pos > 0
-                @pub_year=((match[0,2].to_i) - 1).to_s+'--'
-                return @pub_year
-              end
-            end
-          end
-        end
-        return nil
-      end
-
-      #get a 3 digit year like 965 from the date array
-      def get_three_digit_year dates
-        dates.each do |f_date|
-          matches=f_date.scan(/\d{3}/)
-          if matches.length > 0
-            return matches.first
-          end
-        end
-        return nil
-      end
-      #get the 3 digit BC year, return it as a negative, so -700 for 300 BC. Other methods will translate it to proper display, this is good for sorting.
-      def get_bc_year dates
-        dates.each do |f_date|
-          matches=f_date.scan(/\d{3} B.C./)
-          if matches.length > 0
-            bc_year=matches.first[0..2]
-            return (bc_year.to_i-1000).to_s
-          end
-        end
-        return nil
-      end
-
-      #get a single digit century like '9th century' from the date array
-      def get_single_digit_century dates
-        dates.each do |f_date|
-          matches=f_date.scan(/\d{1}th/)
-          if matches.length == 1
-            @pub_year=((matches.first[0,2].to_i)-1).to_s+'--'
-            return @pub_year
-          end
-          #if there are multiples, check for ones with CE after them
-          if matches.length > 0
-            matches.each do |match|
-              pos = f_date.index(Regexp.new(match+'...CE'))
-              pos = pos ? pos.to_i : f_date.index(Regexp.new(match+' century CE'))
-              pos = pos ? pos.to_i : 0
-              if f_date.include?(match+' CE') or pos > 0
-                @pub_year=((match[0,1].to_i) - 1).to_s+'--'
-                return @pub_year
-              end
-            end
-          end
-        end
-        return nil
-      end
-
-      # @return [Array<String>] dates from dateIssued and dateCreated tags from origin_info with encoding="marc"
-      def dates_marc_encoding
-        @dates_marc_encoding ||= begin
-          parse_dates_from_originInfo
-          @dates_marc_encoding
-        end
-      end
-
-      # @return [Array<String>] dates from dateIssued and dateCreated tags from origin_info with encoding not "marc"
-      def dates_no_marc_encoding
-        @dates_no_marc_encoding ||= begin
-          parse_dates_from_originInfo
-          @dates_no_marc_encoding
-        end
-      end
-
-      # Populate @dates_marc_encoding and @dates_no_marc_encoding from dateIssued and dateCreated tags from origin_info
-      # with and without encoding=marc
-      def parse_dates_from_originInfo
-        @dates_marc_encoding = []
-        @dates_no_marc_encoding = []
-        self.origin_info.dateIssued.each { |di|
-          if di.encoding == "marc"
-            @dates_marc_encoding << di.text
-          else
-            @dates_no_marc_encoding << di.text
-          end
-        }
-        self.origin_info.dateCreated.each { |dc|
-          if dc.encoding == "marc"
-            @dates_marc_encoding << dc.text
-          else
-            @dates_no_marc_encoding << dc.text
-          end
-        }
-      end
     end # class Record
   end # Module Mods
 end # Module Stanford
